Extract factory
