+++ source/Messaging.Infrastructure/Configuration/CompositionRoot.cs
@@ -243,6 +243,7 @@ namespace Messaging.Infrastructure.Configuration
             _services.AddTransient<INotificationHandler<EndOfSupplyNotificationChangedToPending>, NotifyCurrentEnergySupplierWhenConsumerHasMovedIn>();
             _services.AddTransient<INotificationHandler<CustomerMasterDataWasReceived>, SendCustomerMasterDataToEnergySupplierWhenDataIsReceived>();
             _services.AddTransient<INotificationHandler<BusinessProcessWasCompleted>, NotifyGridOperatorWhenConsumerHasMovedIn>();
+            _services.AddTransient<CustomerMasterDataMessageFactory>();
             return this;
         }
